requirements.txt
tensorflow
pandas
Pillow
scikit-learn
tqdm
streamlit
joblib
matplotlib
numpy
wordcloud
